test: 카테고리, 워크스페이스 관련 API 경로 수정에 따른 테스트 수정
